Fill field with autocomplete result before submit
When using the autocomplete call in jQuery, selecting an item with the keyboard
causes the field to be populated with that result. However, clicking one does
not. This meant clicking a result would submit using only the partial text in
the field, rather than the result. This fixes that behavior.