Allow insert of a new hash when changing a password
Previously if a user changed his/her password only an update on
existing rows was performed. In the case where a row was missing
(likely from an incomplete principal creation) this would fail. Now an
insert is performed instead to allow new password algorithms to be
added and gradually rolled out.